machineconfig controller: misc changes to simplify the code
