+++ pkg/operator/controller/machineconfig/types.go
@@ -16,31 +16,6 @@ limitations under the License.
 
 package machineconfigcontroller
 
-import (
-	"sync"
-
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/tools/record"
-
-	"github.com/go-logr/logr"
-
-	"github.com/openshift/node-observability-operator/api/v1alpha1"
-)
-
-// MachineConfigReconciler reconciles a NodeObservabilityMachineConfig object
-type MachineConfigReconciler struct {
-	impl
-	sync.RWMutex
-
-	Log           logr.Logger
-	Scheme        *runtime.Scheme
-	EventRecorder record.EventRecorder
-
-	Node          NodeSyncData
-	MachineConfig MachineConfigSyncData
-	CtrlConfig    *v1alpha1.NodeObservabilityMachineConfig
-}
-
 // NodeSyncData is for storing the state
 // of node operations made for enabling profiling
 type NodeSyncData struct {
